refactor: remove duplicated functions
The Button component's onPress prop is assigned to
onPressToPage(pages.map), which is a function call and not a function definition.

This means that upon rendering, the onPressToPage(pages.map) function is immediately called,
which immediately sets the page value to pages.map. This is why the
value changes on the first run, without pressing the button.